+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Added Storm Spiral Animation macro - no sounds yet
  - Migrated Storm Spiral from Macro to Action triggered from template hook
  - Added sound to Storm Spiral
+ - Cleaned up Rising Hurricane to use some of the targeting utils and removed errors
 
 ## [0.52.0] - 2025-01-03